+++ replicante/src/components/discovery/metrics.rs
@@ -1,4 +1,5 @@
 use prometheus::Counter;
+use prometheus::Gauge;
 use prometheus::Histogram;
 use prometheus::HistogramOpts;
 use prometheus::Opts;
@@ -10,32 +11,25 @@ use slog::Logger;
 lazy_static! {
     pub static ref DISCOVERY_COUNT: Counter = Counter::with_opts(
         Opts::new("replicore_discovery_loops", "Number of discovery runs started")
-    ).expect("Failed to create DISCOVERY_COUNT counter");
+    ).expect("Failed to create DISCOVERY_COUNT");
 
     pub static ref DISCOVERY_DURATION: Histogram = Histogram::with_opts(
         HistogramOpts::new(
             "replicore_discovery_duration",
             "Duration (in seconds) of agent discovery runs"
         ).buckets(vec![0.25, 0.5, 1.0, 2.5, 5.0, 10.0, 20.0, 40.0])
-    ).expect("Failed to create DISCOVERY_DURATION histogram");
+    ).expect("Failed to create DISCOVERY_DURATION");
 
     pub static ref DISCOVERY_ERRORS: Counter = Counter::with_opts(
         Opts::new("replicore_discovery_errors", "Number of errors during agent discovery")
-    ).expect("Failed to create DISCOVERY_ERRORS counter");
+    ).expect("Failed to create DISCOVERY_ERRORS");
 
-    //pub static ref DISCOVERY_PROCESS_ERRORS_COUNT: Counter = Counter::with_opts(
-    //    Opts::new(
-    //        "replicante_discovery_process_errors",
-    //        "Number of errors during processing of discovered agents"
-    //    )
-    //).expect("Failed to create DISCOVERY_PROCESS_ERRORS_COUNT counter");
-
-    //pub static ref DISCOVERY_SNAPSHOT_TRACKER_COUNT: Counter = Counter::with_opts(
-    //    Opts::new(
-    //        "replicante_discovery_snapshot_tracker",
-    //        "Number of clusters tracked by the discovery snapshots emission tracker"
-    //    )
-    //).expect("Failed to create DISCOVERY_SNAPSHOT_TRACKER_COUNT counter");
+    pub static ref DISCOVERY_SNAPSHOT_TRACKED_CLUSTERS: Gauge = Gauge::with_opts(
+        Opts::new(
+            "replicore_discovery_snapshot_tracked_clusters",
+            "Number of clusters tracked by the discovery snapshots emission tracker"
+        )
+    ).expect("Failed to create DISCOVERY_SNAPSHOT_TRACKED_CLUSTERS");
 }
 
 
@@ -52,10 +46,7 @@ pub fn register_metrics(logger: &Logger, registry: &Registry) {
     if let Err(error) = registry.register(Box::new(DISCOVERY_ERRORS.clone())) {
         debug!(logger, "Failed to register DISCOVERY_ERRORS"; "error" => ?error);
     }
-    //if let Err(error) = registry.register(Box::new(DISCOVERY_PROCESS_ERRORS_COUNT.clone())) {
-    //    debug!(logger, "Failed to register DISCOVERY_PROCESS_ERRORS_COUNT"; "error" => ?error);
-    //}
-    //if let Err(error) = registry.register(Box::new(DISCOVERY_SNAPSHOT_TRACKER_COUNT.clone())) {
-    //    debug!(logger, "Failed to register DISCOVERY_SNAPSHOT_TRACKER_COUNT"; "error" => ?error);
-    //}
+    if let Err(error) = registry.register(Box::new(DISCOVERY_SNAPSHOT_TRACKED_CLUSTERS.clone())) {
+        debug!(logger, "Failed to register DISCOVERY_SNAPSHOT_TRACKED_CLUSTERS"; "error" => ?error);
+    }
 }