+++ replicante/src/lib.rs
@@ -65,6 +65,7 @@ pub use self::error::Error;
 pub use self::error::ErrorKind;
 pub use self::error::Result;
 pub use self::tasks::ReplicanteQueues;
+pub use self::tasks::payload as task_payload;
 
 
 lazy_static! {