+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Events streaming interface (backed by store).
 - Fine-grained log level configuration.
 - Grafana annotations backend (through JSON datasource).
+- HA `discovery` component.
 - Periodically emit snapshot events.
 - ROADMAP.md to document "nearby" versions and their "aim".